Avoid segfault in library artwork thread
When changing Library views, the artwork thread may suddenly be trying to
operate with an invalid iterator.  That results in a segfault.

The fix is to have artwork grab the lock before it clears the data away,
and give the lock back just before it's ready to update.

Ridealong is a few places where GtkTreeSelection.unselect_all can replace
the manual equivalent.